requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.1
pandas==2.2.2
python-dotenv==1.0.1
requests==2.32.3
